% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_census_dir}
\alias{check_census_dir}
\title{Helper Function: Check census dir}
\usage{
check_census_dir(.census_geo, .census_dir)
}
\arguments{
\item{.census_geo}{same as in download_census()}

\item{.census_dir}{same as in download_census()}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check census dir
}
