% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_cols}
\alias{check_cols}
\title{Check for Column Names}
\usage{
check_cols(.tab, .require)
}
\arguments{
\item{.tab}{The input table}

\item{.require}{Required column names}
}
\value{
Throws an error if required column is not present in the dataset
}
\description{
Check for Column Names
}
