% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_geo_columns}
\alias{check_geo_columns}
\title{Helper Function: Check if geo columns are present}
\usage{
check_geo_columns(.tab, .census_geo)
}
\arguments{
\item{.tab}{same as in race_...()}

\item{.census_geo}{same as in race_...()}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check if geo columns are present
}
