% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_id}
\alias{check_id}
\title{Helper Function: Check if ID is present in datafram}
\usage{
check_id(.tab)
}
\arguments{
\item{.tab}{same as in race_...()}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check if ID is present in datafram
}
