% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_census}
\alias{download_census}
\title{Download Census Data}
\usage{
download_census(
  .key = "",
  .geo,
  .dir,
  .workers = 1,
  .retry = 10,
  .progress = FALSE
)
}
\arguments{
\item{.key}{A required character object containing a valid Census API key, which can be requested at
\href{https://api.census.gov/data/key_signup.html}{api.census.gov}}

\item{.geo}{Either "county", "tract", "block", or "place"}

\item{.dir}{Folder to store data}

\item{.workers}{Number of paralell workers for downloading data}

\item{.retry}{The number of retries at the census website if network interruption occurs.}

\item{.progress}{Progress Bar}
}
\value{
A Dataframe
}
\description{
Download Census Data
}
