% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lft}
\alias{lft}
\title{List Files in Dataframe}
\usage{
lft(.dirs, .reg = "*", .rec = FALSE)
}
\arguments{
\item{.dirs}{Full paths to folders}

\item{.reg}{Regulare Expression for file names}

\item{.rec}{Recursive search}
}
\value{
A Dataframe
}
\description{
List Files in Dataframe
}
