% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{name_table}
\alias{name_table}
\title{100 First Name/Last Name Pairing with additional Information}
\format{
A data frame with 200 rows and 8 variables:
\describe{
\item{id}{A unique identifier for the individual}
\item{first_name}{First name of the individual}
\item{last_name}{Last name of the individual}
\item{gen}{Gender of the individual. male/female}
\item{age }{Age of the individual}
\item{state}{State of the Individual}
\item{county}{County of the Individual}
\item{tract}{Tract of the Individual}
}
}
\usage{
name_table
}
\description{
A dataset containing Names and other Informations used to predict the race of an individual
}
\keyword{datasets}
