% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_gender}
\alias{predict_gender}
\title{Title}
\usage{
predict_gender(.tab, .use_birth = FALSE, .methods = c("ssa", "ipums", "napp"))
}
\arguments{
\item{.tab}{Input Table (see details)}

\item{.use_birth}{Should race be inferred from information about a persons birth year?}

\item{.methods}{Any combination of "ssa", "ipums", "napp"}
}
\value{
A Dataframe
}
\description{
Title
}
