% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{quiet}
\alias{quiet}
\title{Suppress Messages and Warnings}
\usage{
quiet(..., messages = FALSE, cat = FALSE)
}
\arguments{
\item{...}{Any expression}

\item{messages}{...}

\item{cat}{...}
}
\value{
Any object
}
\description{
Suppress Messages and Warnings
}
