% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{race_predict}
\alias{race_predict}
\title{Predict Race}
\usage{
race_predict(...)
}
\arguments{
\item{...}{Tables generated by race_prr(), race_wru(), or the race_eth python function}
}
\value{
The dataframes in ... are appended and the following columns are added:\cr
\itemize{
\item hp (Highest Probability): Shows the highest probability of the race prediction for each row
\item gd (Guess Difference): Shows the difference between the highest and second highest probability
\item rank_hp: Prediction rank based on the highest probability (hp)
\item rank_gd: Prediction rank based on the guess difference (gd)
}
}
\description{
Predict Race
}
\details{
The column 'method' encodes several variables at one and is encoded in the following way:\cr
{package}-{first name used}-{last name used}-{geo loaction used}-{age used}-{gender used}-{geo location}\cr
For example: The method WRU-0-1-1-1-1-C means that the package WRU is used, prediction is not based on the first name,
predictio is based on the last name, geo location is used, age is used, gender is used, and the geo location is a COunty (C)
}
