% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{race_prr}
\alias{race_prr}
\title{Wrapper around predictrace::predict_race()}
\usage{
race_prr(.tab, .use = c("first_name", "last_name"))
}
\arguments{
\item{.tab}{Input Table (see details)}

\item{.use}{Which name variables to use? First Name (first_name) or Last Name (last_name)}
}
\value{
The original data frame (.tab in long format) appended with the following columns:\cr
\itemize{
\item method: Used methods (here: prr for predictrace)\cr
\item use_name: which name variable has been used (here: first/last)\cr
\item use_geo: Which geo variable has been used (here: none)\cr
\item use_age: Is a persons age used for prediction (here: FALSE)\cr
\item use_sex: Is a persons sex used for prediction (here: FALSE)\cr
\item prob_asian: Probability of the person being asian\cr
\item prob_black: Probability of the person being black\cr
\item prob_hispanic: Probability of the person being hispanic\cr
\item prob_white: Probability of the person being white\cr
\item prob_other: Probability of the person being non of the above races\cr
}

IMPORTANT: Non-Predictions are omited from the output
}
\description{
Wrapper around predictrace::predict_race()
}
\examples{
library(Rrace)

tab_names <- name_table
# Using First/Last Name
race_prr(tab_names)

# Using First Name
race_prr(tab_names, "first_name")

# Using Last Name
race_prr(tab_names, "last_name")

# Using Middle Name (Throws an error, because there's no prediction on middle names)
# If a middle name is present in your dataset, consider pasting it together iwth the first name
\dontrun{
race_prr(tab_names, "middle_name")
}

}
