% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{race_select}
\alias{race_select}
\title{Select Race}
\usage{
race_select(.tab, .col = c("guess_diff", "highest_prob"))
}
\arguments{
\item{.tab}{A dataframe}

\item{.col}{c("guess_diff", "highest_prob")}
}
\value{
A dataframe
}
\description{
Select Race
}
