% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wru_get_combinations}
\alias{wru_get_combinations}
\title{Helper Function: get combinations for race_wru()}
\usage{
wru_get_combinations(.use_age, .use_gen)
}
\arguments{
\item{.use_age}{same as in race_wru()}

\item{.use_gen}{same as in race_wru()}
}
\value{
A datafram
}
\description{
Helper Function: get combinations for race_wru()
}
