% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wru_read_check_census}
\alias{wru_read_check_census}
\title{Helper Function: read and check census files for race_wru()}
\usage{
wru_read_check_census(.tab, .census_dir, .census_geo, .use_age, .use_gen)
}
\arguments{
\item{.tab}{same as in race_wru()}

\item{.census_dir}{same as in race_wru()}

\item{.census_geo}{same as in race_wru()}

\item{.use_age}{same as in race_wru()}

\item{.use_gen}{same as in race_wru()}
}
\value{
A datafram
}
\description{
Helper Function: read and check census files for race_wru()
}
