% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{race_predict}
\alias{race_predict}
\title{Predict Race}
\usage{
race_predict(
  .tab,
  .packages = c("prr", "wru"),
  .prr_use = c("first_name", "last_name"),
  .wru_use_geo = NULL,
  .wru_use_birth = FALSE,
  .wru_use_gender = FALSE,
  .wru_census = NULL
)
}
\arguments{
\item{.tab}{Input Table (see details)}

\item{.packages}{Which package to use, either prr (predictrace) or wru (WRU)}

\item{.prr_use}{Which name variables to use in prr? First Name (first_name) or Last Name (last_name)}

\item{.wru_use_geo}{Which geo variable has been used in wru (Either "county", "tract", "block", or "place").
If used must be the same as in download_census()}

\item{.wru_use_birth}{Should race be inferred from information about a persons birth year (wru)?}

\item{.wru_use_gender}{Should race be inferred from information about a persons sex (wru)?}

\item{.wru_census}{Only needed if .use_geo is not NULL (wru)}
}
\value{
The original data frame (.tab in long format) appended with the following columns:\cr
\itemize{
\item method: Used methods (here: prr for predictrace)\cr
\item use_name: which name variable has been used (here: first/last)\cr
\item use_geo: Which geo variable has been used (here: none)\cr
\item use_birth: Is a persons age used for prediction (here: FALSE)\cr
\item use_gender: Is a persons sex used for prediction (here: FALSE)\cr
\item prob_asian: Probability of the person being asian\cr
\item prob_black: Probability of the person being black\cr
\item prob_hispanic: Probability of the person being hispanic\cr
\item prob_white: Probability of the person being white\cr
\item prob_other: Probability of the person being non of the above races\cr
\item highest_prob: Highest probability of the race predictions
\item guess_diff: difference between the highest and second highest prediction
\item race: the race of the highest prediction
}

IMPORTANT: Non-Predictions are omited from the output
}
\description{
Predict Race
}
