% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{race_wru}
\alias{race_wru}
\title{Wrapper around wru::predict_race()}
\usage{
race_wru(
  .tab,
  .use_geo = NULL,
  .use_age = FALSE,
  .use_sex = FALSE,
  .census = NULL
)
}
\arguments{
\item{.tab}{Input Table (see details)}

\item{.use_geo}{Which geo variable has been used (Either "county", "tract", "block", or "place").
If used must be the same as in download_census()}

\item{.use_age}{Should race be inferred from information about a persons age?}

\item{.use_sex}{Should race be inferred from information about a persons sex?}

\item{.census}{Only needed if .use_geo is not NULL}
}
\value{
The original data frame (.tab in long format) appended with the following columns:\cr
\itemize{
\item method: Used methods (here: prr for predictrace)\cr
\item use_name: which name variable has been used (here: first/last)\cr
\item use_geo: Which geo variable has been used (here: none)\cr
\item use_age: Is a persons age used for prediction (here: FALSE)\cr
\item use_sex: Is a persons sex used for prediction (here: FALSE)\cr
\item prob_asian: Probability of the person being asian\cr
\item prob_black: Probability of the person being black\cr
\item prob_hispanic: Probability of the person being hispanic\cr
\item prob_white: Probability of the person being white\cr
\item prob_other: Probability of the person being non of the above races\cr
}

IMPORTANT: Non-Predictions are omited from the output
}
\description{
Wrapper around wru::predict_race()
}
